Sort & remove misc directory
