+++ .gitignore
@@ -8,6 +8,7 @@
 venv
 *.py[cod]
 *$py.class
+data/
 
 # Protobuf auto-generated bindings
 firmware/src/protocol.pb.c
